// App/src/app/models/Employee.ts
export class Employee {
  id: number;
  nome: string;
  dataDeNascimento: string;
  mae: string;
  pai: string;
  endereco: string;
  cep: Number;
  bairro: string;
  cidade: string;
  uf: string;
  estadoCivil: string;
  grauDeInstrucao: string;
  telefone: Number;
  rg: string;
  cpf: string;
  pis: string;
  dataAdmissao: string;
  funcao: string;
  departamento: string;
  custoDiario: Number;
  escalaDeTrabalho: string;
  ativo: boolean;
}